Struct elma::lev::Level

pub struct Level {
    pub version: Version,
    pub link: i32,
    pub integrity: [f64; 4],
    pub name: CString,
    pub lgr: CString,
    pub ground: CString,
    pub sky: CString,
    pub polygons: Vec<Polygon>,
    pub objects: Vec<Object>,
    pub pictures: Vec<Picture>,
    pub top10_single: Vec<ListEntry>,
    pub top10_multi: Vec<ListEntry>,
    // some fields omitted
}

Level struct that contains all level information.

Fields

version

Elma or Across level.

link

Random number that links level file to replay files.

integrity

Contains four integrity checks (See create_integrity()).

name

Level name.

lgr

LGR file name.

ground

Ground texture name.

sky

Sky texture name.

polygons

Vector with all polygons (See Polygon).

objects

Vector with all objects (See Object).

pictures

Vector with all pictures (See Picture).

top10_single

Vector of Top10 single-player names and times.

top10_multi

Vector of Top10 multi-player names and times.

Methods

impl Level

fn new() -> Level

Returns a new Level struct.

Examples

let level = elma::lev::Level::new();

fn load_level(filename: &str) -> Level

Loads a level file and returns a Level struct.

Examples

let level = elma::lev::Level::load_level("tests/test.lev");

fn get_raw(self) -> Vec<u8>

Converts all struct fields into raw binary form and returns it.

fn save_lev(self, filename: &str)

Saves level as a file.